# Simplify pop logic